+++ Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/Private/Api/AccelByteUserApi.cpp
@@ -43,10 +43,8 @@ void User::LoginWithOtherPlatform(EAccelBytePlatformType PlatformType, const FSt
 
 	if (Credentials.GetSessionState() == Credentials::ESessionState::Valid)
 	{
-		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([this]()
-		{
-			AccelByte::Api::User::Credentials.ForgetAll();
-		}), FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
+		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([](){}), 
+			FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
 		{
 			OnError.ExecuteIfBound(ErrorCode, ErrorMessage);
 		}));
@@ -77,10 +75,8 @@ void User::LoginWithUsername(const FString& Username, const FString& Password, c
 	User::TempUsername = Username;
 	if (Credentials.GetSessionState() == Credentials::ESessionState::Valid)
 	{
-		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([this]()
-		{
-			AccelByte::Api::User::Credentials.ForgetAll();
-		}), FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
+		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([](){}), 
+			FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
 		{
 			OnError.ExecuteIfBound(ErrorCode, ErrorMessage);
 		}));
@@ -110,10 +106,8 @@ void User::LoginWithDeviceId(const FVoidHandler& OnSuccess, const FErrorHandler&
 
 	if (Credentials.GetSessionState() == Credentials::ESessionState::Valid)
 	{
-		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([this]()
-		{
-			AccelByte::Api::User::Credentials.ForgetAll();
-		}), FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
+		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([](){}), 
+			FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
 		{
 			OnError.ExecuteIfBound(ErrorCode, ErrorMessage);
 		}));
@@ -153,10 +147,8 @@ void User::LoginWithLauncher(const FVoidHandler& OnSuccess, const FErrorHandler 
 
 	if (Credentials.GetSessionState() == Credentials::ESessionState::Valid)
 	{
-		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([this]()
-		{
-			AccelByte::Api::User::Credentials.ForgetAll();
-		}), FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
+		Oauth2::Logout(Credentials.GetUserSessionId(), FVoidHandler::CreateLambda([](){}), 
+			FErrorHandler::CreateLambda([OnError](int32 ErrorCode, const FString& ErrorMessage)
 		{
 			OnError.ExecuteIfBound(ErrorCode, ErrorMessage);
 		}));
@@ -551,10 +543,9 @@ void User::GetUserByOtherPlatformUserId(EAccelBytePlatformType PlatformType, con
 	Report report;
 	report.GetFunctionLog(FString(__FUNCTION__));
 	FString PlatformId      = PlatformStrings[static_cast<std::underlying_type<EAccelBytePlatformType>::type>(PlatformType)];
-	TCHAR* UrlFormat        = TEXT("%s/v3/public/namespaces/%s/platforms/%s/users/%s");
 
 	FString Authorization   = FString::Printf(TEXT("Bearer %s"), *Credentials.GetUserSessionId());
-	FString Url             = FString::Printf(UrlFormat, *Settings.IamServerUrl, *Settings.Namespace, *PlatformId, *OtherPlatformUserId);
+	FString Url             = FString::Printf(TEXT("%s/v3/public/namespaces/%s/platforms/%s/users/%s"), *Settings.IamServerUrl, *Settings.Namespace, *PlatformId, *OtherPlatformUserId);
 	FString Verb            = TEXT("GET");
 	FString Accept          = TEXT("application/json");