+++ Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/Public/Core/AccelByteError.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2019 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2018-2020 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
@@ -276,6 +276,7 @@ namespace AccelByte
 		//
 		//Statistic Error Code List
 		//
+		StatisticConfigNotFoundException = 70131,
 		StatisticNotFoundException = 70331,
 		InvalidStatOperatorException = 70330,
 		StatNotDecreasableException = 70334,